She is fuckin done, need more testing though
I am rather proud of this, I made the backasswards mathematical way of applying gravity to sliding and slopes using ternary and conditional operators on the ground and technically wall angles. Simply put, the grav is higher on slopes for that slidey feel, and otherwise normal. The speed of the slide also interpolates differently if on flat ground because it is IMO impossible to do otherwise (I tried in excess of 300 different methods, though all similar, at least 7 to 10 unique thought processes

can you tell I am stoked?